+++ routes/api.php
@@ -11,6 +11,7 @@ use Illuminate\Support\Facades\Route;
 Route::get('/languages', [ReferenceDataController::class, 'languages']);
 Route::get('/translations', [ReferenceDataController::class, 'translations']);
 Route::get('/canons/{canon:code}/books', [ReferenceDataController::class, 'canonBooks']);
+Route::get('/translations/{translationCode}/books', [ReferenceDataController::class, 'translationBooks']);
 Route::get('/translations/{translationCode}/books/{bookSlug}/chapters/{chapter}', [ChapterController::class, 'show'])
     ->whereNumber('chapter');
 Route::get('/strong/{number}', [StudyDataController::class, 'strongEntry']);